move auth.Handler to handler pkg
